test: factor out helpers and spec fixtures
