+++ __tests__/rules/__fixtures__/types.ts
@@ -1,0 +1,6 @@
+export type CommonMarkSpecEntry = {
+    section: string;
+    number: number;
+    markdown: string;
+    html: string;
+};
