+++ jest.config.js
@@ -9,6 +9,7 @@ module.exports = {
     },
     moduleNameMapper: {
         '^src/(.*)': '<rootDir>/src/$1',
+        '^__tests__/(.*)': '<rootDir>/__tests__/$1',
     },
-    testPathIgnorePatterns: [`.*__fixtures__${path.sep}.*`],
+    testPathIgnorePatterns: [`.*__fixtures__${path.sep}.*`, `.*__helpers__${path.sep}.*`],
 };
