+++ tsconfig.json
@@ -12,7 +12,8 @@
     "skipLibCheck": true,
     "baseUrl": ".",
     "paths": {
-      "src/*": ["src/*"]
+      "src/*": ["src/*"],
+      "__tests__/*": ["__tests__/*"]
     }
   }
 }
